OSM File in Conjunction with LiDAR
This update lines up OSM file data with LiDAR data allowing you to use a file as an alternate source to live OSM server data for golf linework in conjunction with LiDAR.  Previously a file could only be used to create a flat course file and did not line up with imported LiDAR.

One other small change included in this update is the option to include LiDAR based trees on background terrain.

This commit includes many changes because I developed it locally before forking the TGC-Designer-Tools project.